Separate dividends where tax has not been withheld
These are supposed to be in Table 8.1, not in Table 8.8.

Also, only return last two years. The first year so it's posssible to
compare in EMTA that the values are still correct (e.g. maybe you forgot
records from one year) and the last year for the new tax report.