Improve Dividend and WithholdingTax matching
In 99.9% of cases the report date is always the same for both records.
However I encountered one case where:
* Dividend A is reported      (report date A)
* Dividend A tax is withheld  (report date A)
* Dividend A tax is sent back (report date B)
* Dividend A tax is withheld  (report date B)

Basically it seemed that there was an error and they fixed the divident
tax amount in a later date. Previously we were comparing `reportDate` to
match divident and withholding_tax records. Instead, we can use
`dateTime` which seems to identify these records better. In the example
above, all of the dateTime's were exactly the same.